tests: Move Namespace functions to testuite
Some functions dealing with namespaces were defined in libnet. However,
they are pretty generic and can be moved to a more common package. This
avoids unneeded dependency: testsuite does not have to import libnet
anymore.
